Added temperature and rain graph with D3.js
The graph is displayed between the temperature and the forecast as an
SVG, and uses the same forecast.io data used for the other weather
information.  It displays the current temperature as a red line and the
chance of rain in filled blue over the next 12 hours.  32 degrees is
marked with a blue line, and 80 degrees with an orange one.

Also cleaned up the CSS a little and added some comments for easier
maintenance.