Significant updates to holidays
Three holidays are now shown at a time (this can be easily changed in
config.js), with each successive holiday being more faded than the last.
The date format now shows the Today, Tomorrow, the day of the upcoming
week, or a date in the format of MM/DD.

More significantly is the addition of holiday filters, in case you don't
want to see some of the more obscure holidays.  You can also add custom
holidays, which show up in gold in the list.