Tweaks to holiday updating
Trying to track down why the holidays aren't updating correctly, which
is difficult since they are only refreshed once a day.  Added a "last
updated" field for the holidays, and adjusted when the timers are fired
to try to make things more robust.  Also re-centered the "summary" text,
since it looked weird justified.