Adds RSS "story dots" and "feels like" current temperature
When the "feels like" apparent temperature is different from the actual
temperature, the phrase "feels like" followed by the apparent
temperature is displayed below the current temperature in smaller text.

At the very bottom of the screen a series of dots indicate which RSS
feed is being viewed, and which story in the feed is currently being
shown.