Added experimental NYC MTA support
Thanks to Ken Balestrieri, the mirror now supports NYC MTA train
schedule alerts.  As with MBTA, you need to get an API key to use it.
You provide the start and end stations, and a time of day that you want
to swap them (ie: going to work vs. going home).  MTA seems to provide a
simple alert string without any useful metadata, so it's just a simple
block of text.

Also updated some of the MBTA functions to make them more useful when
called from other code, like the MTA code.  For example, the array of
alerts and the code to update the div are used by the MTA code, and the
service name is now a variable that can be changed by by the service.