Updated holidays for new holidayapi.com requirements and features
holidayapi.com now requires https, so we do that now.  It also requires
an API key, which you can get for free from that site.

I've also added supoprt for filtering public and/or federal holidays,
and showing the observed date instead of the actual date.  By default
both public and federal holidays are shown, and the actual date is used.
All of these options and the API key are set in the config.