Moved holidays to mm-holidays.js, added internal holiday list
Added mm-holidays.js, which does not rely on external APIs for holidays.
HolidayAPI.com is still available if you don't mind paying $1.50/month
for a key.  The internal list of holidays is now the default, and
currently supports many common US holidays.  New holidays can be added
as per the comments in mm-holidays.js.