Updated MBTA to v3, added CORS option, cleaned up weather PHP calls
MBTA updated to v3, since the old API has been discontinued.  The API key for v3 is optional for small amounts of data, but we use one anyway.  We also support CORS instead of the PHP wrapper, although this requires some server changes.

Small changes to sanatize input in the weather module.  This cannot use CORS (DarkSky specifically disables it so your key doesn't get sniffed), so we continue to use the PHP wrapper here.